@@ -22,39 +22,40 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "15.2.10",
-    "@angular/common": "15.2.10",
-    "@angular/compiler": "15.2.10",
-    "@angular/core": "15.2.10",
-    "@angular/elements": "^15.2.10",
-    "@angular/forms": "15.2.10",
-    "@angular/platform-browser": "15.2.10",
-    "@angular/platform-browser-dynamic": "15.2.10",
-    "@angular/router": "15.2.10",
+    "@angular/animations": "16.2.12",
+    "@angular/common": "16.2.12",
+    "@angular/compiler": "16.2.12",
+    "@angular/core": "16.2.12",
+    "@angular/elements": "^16.2.12",
+    "@angular/forms": "16.2.12",
+    "@angular/platform-browser": "16.2.12",
+    "@angular/platform-browser-dynamic": "16.2.12",
+    "@angular/router": "16.2.12",
     "@babel/runtime": "^7.19.4",
     "@project-sunbird/client-services": "^3.4.8",
     "@project-sunbird/sb-styles": "0.0.7",
-    "@project-sunbird/sunbird-pdf-player-v9": "^4.8.2",
-    "@project-sunbird/sunbird-player-sdk-v9": "^4.6.4",
+    "@project-sunbird/sunbird-pdf-player-v9": "6.0.0",
+    "@project-sunbird/sunbird-player-sdk-v9": "6.0.5",
     "@project-sunbird/telemetry-sdk": "1.3.0",
     "core-js": "^2.5.4",
     "document-register-element": "^1.7.2",
     "reflect-metadata": "^0.1.13",
     "rxjs": "~6.6.7",
     "tslib": "^2.0.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.13.3",
+    "ally.js": "1.4.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "15.2.10",
-    "@angular-devkit/core": "^15.2.10",
-    "@angular-eslint/builder": "15.2.1",
-    "@angular-eslint/eslint-plugin": "15.2.0",
-    "@angular-eslint/eslint-plugin-template": "15.2.1",
-    "@angular-eslint/schematics": "15.2.1",
-    "@angular-eslint/template-parser": "15.2.1",
-    "@angular/cli": "15.2.10",
-    "@angular/compiler-cli": "15.2.10",
-    "@angular/language-service": "15.2.10",
+    "@angular-devkit/build-angular": "16.2.12",
+    "@angular-devkit/core": "^16.2.12",
+    "@angular-eslint/builder": "16.3.1",
+    "@angular-eslint/eslint-plugin": "16.3.1",
+    "@angular-eslint/eslint-plugin-template": "16.3.1",
+    "@angular-eslint/schematics": "16.3.1",
+    "@angular-eslint/template-parser": "16.3.1",
+    "@angular/cli": "16.2.12",
+    "@angular/compiler-cli": "16.2.12",
+    "@angular/language-service": "16.2.12",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.20.15",
@@ -71,7 +72,7 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^15.2.2",
+    "ng-packagr": "^16.2.3",
     "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
     "tsickle": "0.39.1",
