@@ -1,14 +1,14 @@
 {
   "name": "@project-sunbird/sunbird-pdf-player-v9",
-  "version": "6.0.0",
+  "version": "6.1.0",
   "peerDependencies": {
-    "@angular/common": "15.2.10",
-    "@angular/core": "15.2.10",
+    "@angular/common": "16.2.12",
+    "@angular/core": "16.2.12",
     "@project-sunbird/client-services": ">= 3.4.8",
     "@project-sunbird/sb-styles": ">= 0.0.7"
   },
   "dependencies": {
-    "@project-sunbird/sunbird-player-sdk-v9": "5.1.0",
+    "@project-sunbird/sunbird-player-sdk-v9": "6.0.5",
     "tslib": "^2.0.0"
   },
   "homepage": "https://github.com/Sunbird-Knowlg/sunbird-pdf-player/blob/release-6.0.0/README.md",
